Mise en place système d'arrivée. timestamp devient un double

+++ src/fr/ecp/is1220/projet/part1/event_v2/Transportation.java
@@ -44,7 +44,7 @@ public class Transportation extends Event {
 
 		p1.setPatientState(PatientState.INSTALLING);
 		
-		p1.fillRecord(Integer.toString(p1.getPatientRecord().size()) + " - " + Integer.toString(p1.getId()) +" - Transporté par " + transporter.getName() + " :" + transporter.getId() + " sur " + strecher.getId() + " at " + Integer.toString(this.timeStamp)); 
+		p1.fillRecord(Integer.toString(p1.getPatientRecord().size()) + " - " + Integer.toString(p1.getId()) +" - Transporté par " + transporter.getName() + " :" + transporter.getId() + " sur " + strecher.getId() + " at " + Double.toString(this.timeStamp)); 
 		EndEvent e = new EndEvent(this.timeStamp + 3, this.ed, p1); // Cet event a pour fonction de mettre a jour l'état du patient pour simuler le temps passé dans l'évent "arrival"
 		FreeTransporter e2 = new FreeTransporter(this.timeStamp + 5, this.ed, this.transporter); // le transport dure 5 minutes
 		this.ed.addEventInEventQueue(e);
